clients/git-client/pyproject.toml
[project]
name = "git-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "openai>=1.0.0",
    "openai-agents<=0.0.7",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "anyio>=3.7.1",
]